backend/requirements.txt
# Django dependencies
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.2.0

# FastAPI dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-jose==3.3.0
python-multipart==0.0.9
pydantic==2.6.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27

# Utils
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
websockets==12.0
python-socketio==5.11.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Flask dependencies
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.2.0
python-dotenv==1.0.1 